requirements.txt
discord.py<=1.6
pandas<=0.25.3